requirements.txt
requests>=2.28.0
selenium>=4.0.0
webdriver-manager>=4.0.0
pyyaml>=6.0.0
aiohttp>=3.8.0
numpy>=1.21.0